Move common logic to trait
For better maintainablility this commit moves the logic that is common
between the two DateTime-Classes into a trait.
